+++ gamdpy/runtime_actions/trajectory_saver.py
@@ -1,5 +1,6 @@
 import numpy as np
 import numba
+import json
 from numba import cuda, config
 
 from .runtime_action import RuntimeAction
@@ -67,10 +68,11 @@ class TrajectorySaver(RuntimeAction):
                 chunks=(1, 1, self.configuration.N, self.configuration.D),
                 dtype=np.int32,  compression=self.compression, compression_opts=self.compression_opts)
         output['trajectory_saver'].attrs['compression_info'] = f"{self.compression} with opts {self.compression_opts}"
-        output['trajectory_saver'].attrs['steps'] = self.scheduler.steps
         output['trajectory_saver'].attrs['scheduler'] = self.scheduler.__class__.__name__
+        output['trajectory_saver'].attrs['scheduler_info'] = json.dumps(self.scheduler.kwargs)
         output['trajectory_saver'].attrs['num_timeblocks'] = self.num_timeblocks
         output['trajectory_saver'].attrs['steps_per_timeblock'] = self.steps_per_timeblock
+        # output['trajectory_saver'].create_dataset('steps', data=self.scheduler.steps)
 
         #output.attrs['vectors_names'] = list(self.sid.keys())
         if self.include_simbox: