+++ CHANGELOG.md
@@ -6,8 +6,11 @@
 * Integrator for Brownian dynamics.
 * Integrator for gradient descent.
 * Integrator for NVU dynamics.
+* Variable strain rate for SLLOD
+* calc_dynamics can handle Lees-Edwards boundary conditions
+* Tabulated pair potentials.
 
 ### Bug fixes
-
+* Incorrect application of shifted-force cutoff
 ## Version 0.8.1
 First release of the package.